integration: split up data and fixtures
